command refactor
